requirements.txt
tensorflow==2.5.0
numpy==1.19.2
Werkzeug==1.0.1
keras_nightly==2.5.0.dev2021032900
Flask==1.1.2
seaborn==0.11.0
pandas==1.1.3
keras==2.4.3
Pillow==8.2.0
